Add functionality around source and destination
For now, this just locks the source and destination folders to values
("content" and "public", respectively). But, it also adds tests around
the behavior: no source should cause a panic, and no destination will
create the destination directory.

To be honest, the panic isn't tested. I'm not exactly sure how to
achieve that. But I did pull the work out into a function that can be
tested to ensure that a destination directory is created, etc.

Additionally, I added two tasks (with explanations) to the README, with
one of them being completed (which you're reading about).