Add a factory for creating processors
To be honest, I'm not exactly sure this is idiomatic Rust. But the goal
is to create processors that will then handle the nitty-gritty of
turning the contents of a file (whatever the format) into something
suitable for delivery over the web (in this case HTML or CSS).

What perplexed me was how to reduce some of the boilerplate, but I was
unable to do so. For starters, the `Processor` trait could not have a
field, so each implementation must handle the `contents` individually. I
started going down the path of another trait that did something similar
to the `kind()` quasi-field. But, to be honest, I need to get this in.

Additionally, I don't know how I feel abou the mutable self in the
signature of `write_to(&mut self, path: &Path)`. It seems like
information that no one but the implementor should know.

I am pleased with the tests that I have in place, and continue to drive
the design with them. Obviously, that is the only interesting thing at
this point, because this project doesn't yet do anything.